test-configs.yaml: choose the debian version depending on the arch
In order to test kselftest and ltp on riscv platforms, we currently
have to use debian-sid instead of debian bullseye. The reason for
that is that bullseye does not support riscv platform and only
sid does.
As a consequence, and in order to not duplicate the test plans in
test-configs.yaml, the logic to be able to choose the debian
version depending on the arch that is being tested has been added.
To do so:
- the mention to bullseye has been removed in rootfs in the test
plans defintions
- a image-distros.yaml file has been added and it contains the
information to get the rootfs name with the debian version from
the name (without the debian version, now used in the test plan)
and the architecture.
- a Distro object has been added to contain the information from
the image-distros.yaml
- the TestPlan object now has a Distro object that can contain
different RootFS objects depending on the architecture instead of
a single RootFS object.
- When generating a test, the architecture is known and the correct
RootFs object is chosen.

Signed-off-by: Nicolas Belin <nbelin@baylibre.com>